The main thing here is removal of unnecessary data juggling around namespace paths (before this PR, they are converted from a list to a map to a list of strings to a string). A few other minor things here and there.

Seeing the following improvement using `BenchmarkExecTrue`:
```
goos: linux
goarch: amd64
pkg: github.com/opencontainers/runc/libcontainer/integration
cpu: 12th Gen Intel(R) Core(TM) i7-12800H
            │  before-ns  │             after-ns              │
            │   sec/op    │   sec/op     vs base              │
ExecTrue-20   3.972m ± 2%   3.697m ± 1%  -6.94% (p=0.000 n=8)

            │  before-ns   │              after-ns               │
            │     B/op     │     B/op      vs base               │
ExecTrue-20   26.43Ki ± 0%   21.53Ki ± 1%  -18.54% (p=0.000 n=8)

            │ before-ns  │             after-ns              │
            │ allocs/op  │ allocs/op   vs base               │
ExecTrue-20   244.0 ± 0%   172.0 ± 0%  -29.51% (p=0.000 n=8)
```